requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloud-sql-python-connector==1.20.0
cryptography==46.0.5
decorator==5.2.1
dnspython==2.8.0
Flask==3.1.2
frozenlist==1.8.0
google==3.0.0
google-auth==2.48.0
greenlet==3.3.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
loguru==0.7.3
MarkupSafe==3.0.3
multidict==6.7.1
numpy==2.2.6
pandas==2.3.3
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
rsa==4.9.1
scrapfly-sdk==0.8.24
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
yarl==1.22.0
